Added withdrawal module

+++ BoradbandZone_DbScripts/BroadbandZone_DbScripts/dbo/User Defined Types/udt_CommissionSetting.sql
@@ -1,6 +1,8 @@
 ﻿CREATE TYPE [dbo].[udt_CommissionSetting] AS TABLE (
     [CategoryId]         INT          NULL,
     [Category]           VARCHAR (50) NULL,
-    [SupCommission]      MONEY        NULL,
+    [SupCommission]      SMALLINT     NULL,
     [AgentCommissionPer] SMALLINT     NULL);
+
+
